Renames utils file
